+++ src/test/tests.ts
@@ -4,4 +4,5 @@ import './testSchemaGenerator';
 import './testLogger';
 import './testMocking';
 import './testResolution';
+import './testMakeRemoteExecutableSchema';
 import './testMergeSchemas';